[2Week_Mission] 한승연 - 상품, 장바구니, 주문/결제 기능 리팩토링

+++ 2Week_Mission/mutbooks/src/main/java/com/example/mutbooks/app/member/service/MemberService.java
@@ -117,7 +117,7 @@ public class MemberService {
     }
 
     // 회원의 남은 예치금 잔액 조회
-    public long getRestCash(Member member) {
+    public int getRestCash(Member member) {
         return findByUsername(member.getUsername()).getRestCash();
     }
